Changed File Names to use the more common casing for React-Native
Removed duplicate Screens.
