+++ frontend/src/app/auth/RegisterScreen.tsx
@@ -1,12 +1,29 @@
 import React, {useState} from 'react';
-import { router } from "expo-router";
 import { StyleSheet, SafeAreaView, View, Image, Text, TouchableOpacity, TextInput, ScrollView, ImageBackground, Pressable } from 'react-native';
+import { router } from "expo-router";
+import {Dropdown} from 'react-native-element-dropdown';
 
-export default function Example() {
+const genderData = [
+    {label: 'Male', value: 'male'},
+    {label: 'Female', value: 'female'},
+    {label: 'Non-binary', value: 'Non-binary'},
+    {label: 'Other', value: 'Other'},
+    {label: 'Decline to state', value: 'Decline to state'},
+];
+
+export default function RegisterScreen() {
     const [form, setForm] = useState({
+        firstName: '',
+        lastName: '',
         email: '',
+        confirmEmail: '',
         password: '',
+        confirmPassword: '',
+        gender: '',
     });
+
+
+    const [isFocus, setIsFocus] = useState(false);
 
     return (
         <SafeAreaView style={{flex: 1, backgroundColor: 'black'}}>
@@ -22,21 +39,47 @@ export default function Example() {
                                 style={styles.headerImg}
                                 source={require('@assets/images/Picture 1.png')}
                             />
-
                             <Text style={styles.title}>
-                                Login <Text style={{color: '#ec5707'}}>Thrift Market</Text>
+                                Sign Up <Text style={{color: '#ec5707'}}>Thrift Market</Text>
                             </Text>
                             <Text style={styles.subtitle}>
-                                Get started with our app,just create an account and enjoy the experience
+                                Create an account to enjoy our services
                             </Text>
+
+                            <Text>Sign Up</Text>
+                            <TextInput placeholder="Email" />
+                            <TextInput placeholder="Password" secureTextEntry />
                         </View>
                         <View style={styles.form}>
+                            <View style={styles.input}>
+                                <Text style={styles.inputLabel}>First Name</Text>
+                                <TextInput
+                                    autoCapitalize="none"
+                                    autoCorrect={false}
+                                    onChangeText={(firstName) => setForm({...form, firstName})}
+                                    placeholder="John"
+                                    placeholderTextColor="#6b7280"
+                                    style={styles.inputControl}
+                                    value={form.firstName}
+                                />
+                            </View>
+                            <View style={styles.input}>
+                                <Text style={styles.inputLabel}>Last Name</Text>
+                                <TextInput
+                                    autoCapitalize="none"
+                                    autoCorrect={false}
+                                    onChangeText={(lastName) => setForm({...form, lastName})}
+                                    placeholder="Doe"
+                                    placeholderTextColor="#6b7280"
+                                    style={styles.inputControl}
+                                    value={form.lastName}
+                                />
+                            </View>
                             <View style={styles.input}>
                                 <Text style={styles.inputLabel}>Email address</Text>
                                 <TextInput
                                     autoCapitalize="none"
                                     autoCorrect={false}
-                                    clearButtonMode="while-editing"
                                     keyboardType="email-address"
                                     onChangeText={(email) => setForm({...form, email})}
                                     placeholder="john@example.com"
@@ -46,10 +89,22 @@ export default function Example() {
                                 />
                             </View>
                             <View style={styles.input}>
+                                <Text style={styles.inputLabel}>Confirm Email</Text>
+                                <TextInput
+                                    autoCapitalize="none"
+                                    autoCorrect={false}
+                                    keyboardType="email-address"
+                                    onChangeText={(confirmEmail) => setForm({...form, confirmEmail})}
+                                    placeholder="john@example.com"
+                                    placeholderTextColor="#6b7280"
+                                    style={styles.inputControl}
+                                    value={form.confirmEmail}
+                                />
+                            </View>
+                            <View style={styles.input}>
                                 <Text style={styles.inputLabel}>Password</Text>
                                 <TextInput
                                     autoCorrect={false}
-                                    clearButtonMode="while-editing"
                                     onChangeText={(password) => setForm({...form, password})}
                                     placeholder="********"
                                     placeholderTextColor="#6b7280"
@@ -58,48 +113,81 @@ export default function Example() {
                                     value={form.password}
                                 />
                             </View>
+                            <View style={styles.input}>
+                                <Text style={styles.inputLabel}>Confirm Password</Text>
+                                <TextInput
+                                    autoCorrect={false}
+                                    onChangeText={(confirmPassword) => setForm({...form, confirmPassword})}
+                                    placeholder="********"
+                                    placeholderTextColor="#6b7280"
+                                    style={styles.inputControl}
+                                    secureTextEntry={true}
+                                    value={form.confirmPassword}
+                                />
+                            </View>
+
+                            {/* Gender Dropdown */}
+                            <View style={styles.input}>
+                                <Text style={styles.inputLabel}>Gender</Text>
+                                <Dropdown
+                                    style={[styles.dropdown, isFocus && {borderColor: 'blue'}]}
+                                    placeholderStyle={styles.placeholderStyle}
+                                    selectedTextStyle={styles.selectedTextStyle}
+                                    data={genderData}
+                                    labelField="label"
+                                    valueField="value"
+                                    placeholder={!isFocus ? 'Select gender' : '...'}
+                                    value={form.gender}
+                                    onFocus={() => setIsFocus(true)}
+                                    onBlur={() => setIsFocus(false)}
+                                    onChange={item => {
+                                        setForm({...form, gender: item.value});
+                                        setIsFocus(false);
+                                    }}
+                                />
+                            </View>
+
                             <View style={styles.formAction}>
                                 <TouchableOpacity
                                     onPress={() => {
-                                        // handle onPress
+                                        // handle sign-up logic
                                     }}
                                 >
                                     <View style={styles.btn}>
-                                        <Text style={styles.btnText}>Sign in</Text>
+                                        <Text style={styles.btnText}>Sign Up</Text>
                                     </View>
                                 </TouchableOpacity>
                             </View>
 
-                            {/* Google Sign-In Button */}
+                            {/* Google Sign-Up Button */}
                             <TouchableOpacity
                                 onPress={() => {
-                                    // handle Google sign-in
+                                    // handle Google sign-up logic
                                 }}
                                 style={styles.googleBtn}
                             >
                                 <Image
+                                    style={styles.googleLogo}
                                     //Todo - Change to google_logo.svg
                                     source={require('@assets/images/google_logo.png')}
-                                    style={styles.googleLogo}
                                 />
-                                <Text style={styles.btnText}>Sign in with Google</Text>
+                                <Text style={styles.btnText}>Sign up with Google</Text>
                             </TouchableOpacity>
 
-                            <Text style={styles.formLink}>Forgot password?</Text>
                         </View>
                     </ScrollView>
 
                     <Pressable
                         onPress={() =>
                             router.replace({
-                                pathname: "/auth/register"
+                                pathname: "/auth/LoginScreen"
                             })
                         }
                         style={{marginTop: 'auto'}}
                     >
                         <Text style={styles.formFooter}>
-                            Don't have an account?{' '}
-                            <Text style={{textDecorationLine: 'underline'}}>Sign up</Text>
+                            Already have an account?{' '}
+                            <Text style={{textDecorationLine: 'underline'}}>Sign in</Text>
                         </Text>
                     </Pressable>
                 </View>
@@ -162,12 +250,6 @@ const styles = StyleSheet.create({
         marginTop: 4,
         marginBottom: 16,
     },
-    formLink: {
-        fontSize: 16,
-        fontWeight: '600',
-        color: '#ec5707',
-        textAlign: 'center',
-    },
     formFooter: {
         fontSize: 15,
         fontWeight: '600',
@@ -196,6 +278,20 @@ const styles = StyleSheet.create({
         borderColor: '#C9D3DB',
         borderStyle: 'solid',
     },
+    dropdown: {
+        height: 50,
+        borderColor: 'gray',
+        borderWidth: 0.5,
+        borderRadius: 8,
+        paddingHorizontal: 8,
+        backgroundColor: 'white',
+    },
+    placeholderStyle: {
+        fontSize: 16,
+    },
+    selectedTextStyle: {
+        fontSize: 16,
+    },
     btn: {
         flexDirection: 'row',
         alignItems: 'center',
@@ -207,6 +303,11 @@ const styles = StyleSheet.create({
         backgroundColor: '#ec5707',
         borderColor: '#ec5707',
     },
+    btnText: {
+        fontSize: 18,
+        lineHeight: 26,
+        color: '#ffffff',
+    },
     googleBtn: {
         flexDirection: 'row',
         alignItems: 'center',
@@ -215,21 +316,14 @@ const styles = StyleSheet.create({
         paddingVertical: 10,
         paddingHorizontal: 20,
         borderWidth: 1,
-        backgroundColor: 'rgb(57,56,56)',
-        borderColor: 'rgb(57,56,56)',
-        marginTop: 10,
-        marginBottom: 10
+        backgroundColor: 'rgb(57,56,56)', // Background color for the Google button
+        borderColor: 'rgb(57,56,56)', // Border color for the Google button
+        marginTop: 10, // Add margin to separate from the Sign up button
     },
     googleLogo: {
         width: 24,
         height: 24,
-        marginRight: 8,
-        opacity: 0.8,
-    },
-    btnText: {
-        fontSize: 18,
-        lineHeight: 26,
-        fontWeight: '600',
-        color: '#fff',
+        marginRight: 10,
     },
 });
+
